+++ server/api/views/assistant/assistant_services.py
@@ -4,11 +4,8 @@ import logging
 from openai import OpenAI
 
 from .assistant_prompts import INSTRUCTIONS
-from .tool_services import (
-    SEARCH_TOOLS_SCHEMA,
-    make_search_tool_mapping,
-    handle_tool_calls_with_reasoning,
-)
+from .tool_services import get_tools_schema, make_tool_mapping
+from .agentic_loop import handle_tool_calls_with_reasoning
 
 logger = logging.getLogger(__name__)
 
@@ -44,14 +41,17 @@ def run_assistant(
         "model": "gpt-5-nano",  # 400,000 token context window
         # A summary of the reasoning performed by the model. This can be useful for debugging and understanding the model's reasoning process.
         "reasoning": {"effort": "low", "summary": None},
-        "tools": SEARCH_TOOLS_SCHEMA,
+        # get_tools_schema() / make_tool_mapping() are aggregators owned by
+        # tool_services.py — we deliberately call them here instead of naming individual
+        # tools, so adding a tool never requires editing this file.
+        "tools": get_tools_schema(),
     }
 
-    # TOOLS_SCHEMA tells the model what tools exist and what arguments to generate.
-    # tool_mapping wires those tool names to the Python functions that execute them.
-    # They are separate because the model generates arguments (schema concern) but
-    # cannot supply request-time values like user (mapping concern).
-    tool_mapping = make_search_tool_mapping(user)
+    # The schema (get_tools_schema) tells the model what tools exist and what arguments
+    # to generate; the tool_mapping wires those tool names to the Python functions that
+    # execute them. They are kept separate because the model generates arguments (schema
+    # concern) but cannot supply request-time values like `user` (mapping concern).
+    tool_mapping = make_tool_mapping(user)
 
     if not previous_response_id:
         response = client.responses.create(
